requirements.txt
streamlit
openai
google-genai
azure-identity
requests
python-dotenv